requirements.txt

flask
python-dotenv